change files to be consistent with splunk_inputs
